new mutator + reorganized code
